Fixed a logical error
I fixed a logical error in move_to_target_with_limits(). The issue was that I would change the angle by the minimum, which was 0. This means I would see nothing. I fixed the issue by finding the amount the arm would have to rotate to stay at the angle. I have not been able to test the code yet.