Mr. Steeltastic Arm Tuning (#14)
* Added (Broken) moveToPosPriority

* Added motorToPos

* More arm tweaks

For some reason putting the arm in the pickup pose, then to home, then back to pickup pose doesn't move the top motor at all. This is so much fun. ;-;

* Minor Refactoring

* Added synchronized motor movement

* Made PickupCube faster

and re-added PlaceCubeMid for future use (will need tuning)

* Mega Code Cleanup

Also removed numpy dependencies in arm.py with math, so hopefully that helps with file size on the RoboRIO

* Forgot to remove an import :p

* Added ToggleArmCoast

In case we need to get new positions for the motors (e.g. new CubePlaceMid positions)

* Added toPos function into ArmMotor

* Added generic SetPosition for motors

* Giga Code Rewrite

Moved a lot of things around, changed the position of CubePlaceMid, and streamlined how the ArmCombo movement commands work (replaced the commands with parallel command groups, huge optimization)

* No Progress Today ;-;

Just some minor changes with constants

* Removed unused command

* Minor refactoring

* Major Code Refactoring + New Tick command

Mainly cleanup. The tick command just updates wpilib every so often, saving clutter with a million overlapping smart dashboard calls :p

* Updated default command

* Removed some old comments

* Sim stuff

We really need to add this to the .gitignore its so annoying ;-;

* Removed all simgui files + gitignore

* Forgor these ones

---------

Co-authored-by: Kenneth Heyen <kenneth.heyen@voltacharging.com>